Test files for bug 481.
There are 4 types of meshes that attempt to convert the shading object into
a stream. According to spec (section 8.7.4.1 table 76), the shading object can
be a stream or a dictionary.

All of this shading code assumes it can load the mesh from a stream. The
original code just early exited if it was not a stream. We skip the call
if it is not a stream.

BUG=pdfium:481
R=tsepez@chromium.org

Review URL: https://codereview.chromium.org/1905953002 .