Update text markup annotation expected results to use /QuadPoints when appropriate
This CL is from Antonio Gomes.

In [1], PDFium is being updated to use /QuadPoints
coordinates rather than /Rect coordinates for drawing
text markup annotations, in accordance to the presence of /AP.

This CL updates the expected results for some test cases.

Notice that for the test cases updated here, the difference
is in pixel-level, and not able to see by naked eyes.

[1] https://codereview.chromium.org/2289293005/

BUG=pdfium:585

Review URL: https://codereview.chromium.org/2340933003 .