Update the expected result for a corpus test.
The PDF rendering process was modified to calculate zero area and draw
thin line for each sub path individually instead of for a whole path.[1]
This will result in more accurate rendering results since the properties
of a sub path, such as whether it has an identical matrix or the zero
area should be rendered as thin line, will only be applied to the sub
path itself instead of affecting the whole path.

This CL updates the corpus test result for the new rendering process.

[1]https://pdfium-review.googlesource.com/78030

Bug: pdfium:1638
Change-Id: Ifce7d6f911309b3bda87b283260198dc2df2a651
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium_tests/+/78010
Reviewed-by: Daniel Hosseinian <dhoss@chromium.org>