Update corpus test expectations due to a freetype change.
FreeType change [1] improves font rendering on quadratic bezier arcs
when SSE2 is available or when using 64-bit CPUs.

Since [2] has enabled SSE2 on all MSVC bots and the rest of PDFium bots
are using 64-bit CPU, this CL updates the expectations of corpus tests
which are affected by the FreeType change.

[1] https://crrev.com/56cc2ad446e920164e823e78a21972cadb339bfc
[2] https://pdfium-review.googlesource.com/c/pdfium/+/84433

Change-Id: Ie626d51e7a44141d68614c8d058ca7ef917e1805
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium_tests/+/84291
Reviewed-by: Lei Zhang <thestig@chromium.org>