Update Skia expectation for 11.pdf
The Skia rendering behavior is correct for this test, which is drawing
black hairlines and then drawing white skinny images on top of them
inside the circle path.

There is no rule that the hairlines must be covered by the white
images and the rendering quality is controlled by how Skia handles
extremely "thin" image rendering internally. Therefore, update the
Skia expectation.

Bug: pdfium:1811
Change-Id: I825eb75aa4fd03765549385a4c71fbdae229b726
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium_tests/+/103955
Reviewed-by: K. Moon <kmoon@chromium.org>