Add Skia expectations for xfermodes{2,3}.pdf
Adds Skia expectations for xfermodes2.pdf and xfermodes3.pdf. These
could not be added until we added support for fuzzy matching in
pdfium:1991, as the ColorBurn results are device-specific.

Note that the Skia result for xfermodes2.pdf is now the default
expectation, as it matches Acrobat better than AGG does. In particular,
AGG renders ColorDodge with too much red, and ColorBurn with too little
green.

Bug: pdfium:1959, pdfium:1990
Change-Id: I31b69f7c90dcb6df84a799006ebbbc0cde726b2e
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium_tests/+/103953
Reviewed-by: Nigi <nigi@chromium.org>